fix : Tooltip 오류 수정
